Refactored RequestProtocolVerifier
